requirements.txt
assemblyai
flask
google-generativeai
pytube
pyttsx3